refactor: extract sgid controller
