--- package.json
@@ -1,7 +1,0 @@
-{
-  "dependencies": {
-    "pg": "^8.11.0",
-    "pg-hstore": "^2.3.4",
-    "sequelize": "^6.32.1"
-  }
-}
